move kotsadm join token response type to 'kinds' package
