+++ kinds/types/join/testdata/join-command-response-empty-overrides.yaml
