+++ kinds/types/join/testdata/join-command-response-no-overrides.yaml
