+++ kinds/types/join/testdata/join-command-response-storage-and-port-overrides.yaml
