+++ pkg/kotsadm/client.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"net/http"
 	"time"
+
+	"github.com/replicatedhq/embedded-cluster/kinds/types/join"
 )
 
 var _ ClientInterface = (*Client)(nil)
@@ -15,7 +17,7 @@ type Client struct{}
 
 // GetJoinToken issues a request to the kots api to get the actual join command
 // based on the short token provided by the user.
-func (c *Client) GetJoinToken(ctx context.Context, baseURL, shortToken string) (*JoinCommandResponse, error) {
+func (c *Client) GetJoinToken(ctx context.Context, baseURL, shortToken string) (*join.JoinCommandResponse, error) {
 	url := fmt.Sprintf("https://%s/api/v1/embedded-cluster/join?token=%s", baseURL, shortToken)
 	ctx, cancel := context.WithTimeout(ctx, time.Minute)
 	defer cancel()
@@ -34,7 +36,7 @@ func (c *Client) GetJoinToken(ctx context.Context, baseURL, shortToken string) (
 	if resp.StatusCode != http.StatusOK {
 		return nil, fmt.Errorf("unexpected status code: %d", resp.StatusCode)
 	}
-	var command JoinCommandResponse
+	var command join.JoinCommandResponse
 	if err := json.NewDecoder(resp.Body).Decode(&command); err != nil {
 		return nil, fmt.Errorf("unable to decode response: %w", err)
 	}
